define executeLaunch as an async executor
